fixes to enemy ai

+++ UnityFolder/Shovel Jam Tri/Assets/Scripts/CameraFollow.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class CameraFollow : MonoBehaviour {
-    public int xOffset = -16;
+    public float xOffset = -16;
     public int zOffset = 19;
     public Transform target;
     Vector3 targetPos, smoothedPos;
